+++ src/vector3.cpp
@@ -278,11 +278,11 @@ Vector3 Vector3::Slerp(const Vector3 &v1, const Vector3 &v2, float d) {
     // Acos(dot) returns the angle between start and end,
     // And multiplying that by percent returns the angle between
     // start and the final result.
-    float theta = acos(dot)*d;
+    float theta = acosf(dot)*d;
     Vector3 RelativeVec = v2 - v1*dot;
     RelativeVec.normalize();     // Orthonormal basis
     // The final result.
-    return ((v1*cos(theta)) + (RelativeVec*sin(theta)));
+    return ((v1*cosf(theta)) + (RelativeVec*sinf(theta)));
 }
 
 void Vector3::OrthoNormalize(Vector3 &normal, Vector3 &tangent) {
